Add permissions for build job in CI/CD config
This update to the `quicklypassword-ci-cd.yml` file introduces a `permissions` section under the `build` job. The `checks` permission is set to `write`, and the `contents` permission is set to `read`, enabling the job to write checks and read contents during execution, which is essential for the build and test processes.